Allow user to define puzzles (MVP) (#21)
Made the Sudoku board editable by changing `td` to `input`.

Added basic input validation; user can only enter 1 to 9 and can't enter values which would result in an invalid Sudoku. However, the reasons are only logged to console at the moment. In the future it would be nice if offending cells were highlighted for example.

Also did some refactoring and fixed a bug where the reset button didn't clear cells with `initiallySet = false` (reimplemented the `onClick` code for this).

Closes #3.